+++ src/main/kotlin/br/com/tagliaferrodev/samplerest/repositories/PaisRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PaisRepository : JpaRepository<Pais, Long>
+interface PaisRepository : JpaRepository<Pais, Int>
