+++ src/main/kotlin/br/com/tagliaferrodev/samplerest/repositories/PessoaRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PessoaRepository : JpaRepository<Pessoa, Long>
+interface PessoaRepository : JpaRepository<Pessoa, Int>
